+++ src/renderer/components/Sidebar.tsx
@@ -1,5 +1,6 @@
 import { History, HomeRounded, Menu } from '@mui/icons-material';
 import {
+  Avatar,
   Box,
   Drawer,
   IconButton,
@@ -7,11 +8,12 @@ import {
   ListItem,
   ListItemButton,
   ListItemContent,
+  Stack,
   Typography,
 } from '@mui/joy';
 import { FC, ReactNode } from 'react';
 import { Link, useLocation } from 'react-router-dom';
-
+import logoSrc from '../../../assets/icon.png';
 export interface SidebarProps {
   isOpen: boolean;
   onClose: () => void;
@@ -22,18 +24,8 @@ export const Sidebar: FC<SidebarProps> = ({ isOpen, onClose }) => {
     <Drawer open={isOpen} size="sm" onClose={onClose}>
       <SidebarHeader />
       <List sx={{ p: 2 }}>
-        <SidebarItem
-          text="Inicio"
-          icon={<HomeRounded />}
-          link="/"
-          onClose={onClose}
-        />
-        <SidebarItem
-          text="Historico"
-          icon={<History />}
-          link="/history"
-          onClose={onClose}
-        />
+        <SidebarItem text="Inicio" icon={<HomeRounded />} link="/" onClose={onClose} />
+        <SidebarItem text="Historico" icon={<History />} link="/history" onClose={onClose} />
       </List>
     </Drawer>
   );
@@ -48,10 +40,10 @@ const SidebarHeader: FC = () => (
         alignItems: 'center',
       }}
     >
-      <IconButton variant="soft" color="primary" size="md">
-        <Menu />
-      </IconButton>
-      <Typography level="h3">Menu</Typography>
+      <Stack direction="row" sx={{ ml: 1, mt: 3 }} gap={1.5}>
+        <Avatar src={logoSrc} alt="Sanita Ticket Logo" size="sm" />
+        <Typography level="h3">Sanita Ticket</Typography>
+      </Stack>
     </Box>
   </Box>
 );
@@ -62,14 +54,8 @@ const SidebarItem: FC<{
   link: string;
   onClose: () => void;
 }> = ({ text, icon, link, onClose }) => (
-  <ListItem
-    sx={{ my: 0.5, '--ListItem-radius': (theme) => theme.vars.radius.sm }}
-  >
-    <ListItemButton
-      sx={{ gap: 2 }}
-      selected={useLocation().pathname === link}
-      onClick={onClose}
-    >
+  <ListItem sx={{ my: 0.5, '--ListItem-radius': (theme) => theme.vars.radius.sm }}>
+    <ListItemButton sx={{ gap: 2 }} selected={useLocation().pathname === link} onClick={onClose}>
       {icon}
       <ListItemContent>
         <Link to={link} style={{ textDecoration: 'none' }}>